imlemented some frontend graph loading infrastructure and onclick events of buttons

+++ web/backend/src/main/java/de/exbio/reposcapeweb/db/repositories/nodes/DrugRepository.java
@@ -8,7 +8,7 @@ import java.util.Collection;
 import java.util.List;
 
 @Repository
-public interface DrugRepository extends CrudRepository<Drug,Long> {
+public interface DrugRepository extends CrudRepository<Drug,Integer> {
     List<Drug> findAllByPrimaryDomainIdIn(Collection<String> ids);
 
 }
